+++ fdbclient/AuditUtils.actor.cpp
@@ -29,6 +29,247 @@
 
 #include "flow/actorcompiler.h" // has to be last include
 
+// Check if any pair of ranges are exclusive with each other
+// This is not a part in consistency check of audit metadata
+// This is used for checking the validity of inputs to rangesSame()
+bool elementsAreExclusiveWithEachOther(std::vector<KeyRange> ranges) {
+	ASSERT(std::is_sorted(ranges.begin(), ranges.end(), KeyRangeRef::ArbitraryOrder()));
+	for (int i = 0; i < ranges.size() - 1; ++i) {
+		if (ranges[i].end > ranges[i + 1].begin) {
+			TraceEvent(SevError, "ElementsAreNotExclusiveWithEachOther").detail("Ranges", describe(ranges));
+			return false;
+		}
+	}
+	return true;
+}
+
+// Check if any range is empty in the given list of ranges
+// This is not a part in consistency check of audit metadata
+// This is used for checking the validity of inputs to rangesSame()
+bool noEmptyRangeInRanges(std::vector<KeyRange> ranges) {
+	for (const auto& range : ranges) {
+		if (range.empty()) {
+			return false;
+		}
+	}
+	return true;
+}
+
+// Given a list of ranges, where ranges can overlap with each other
+// Return a list of exclusive ranges which covers the ranges exactly
+// the same as the input list of ranges
+std::vector<KeyRange> coalesceRangeList(std::vector<KeyRange> ranges) {
+	std::sort(ranges.begin(), ranges.end(), [](KeyRange a, KeyRange b) { return a.begin < b.begin; });
+	std::vector<KeyRange> res;
+	for (const auto& range : ranges) {
+		if (res.empty()) {
+			res.push_back(range);
+			continue;
+		}
+		if (range.begin <= res.back().end) {
+			if (range.end > res.back().end) { // update res.back if current range extends the back range
+				KeyRange newBack = Standalone(KeyRangeRef(res.back().begin, range.end));
+				res.pop_back();
+				res.push_back(newBack);
+			}
+		} else {
+			res.push_back(range);
+		}
+	}
+	return res;
+}
+
+// Given two lists of ranges --- rangesA and rangesB, check if two lists are identical
+// If not, return the mismatched two ranges of rangeA and rangeB respectively
+Optional<std::pair<KeyRange, KeyRange>> rangesSame(std::vector<KeyRange> rangesA, std::vector<KeyRange> rangesB) {
+	if (g_network->isSimulated()) {
+		ASSERT(noEmptyRangeInRanges(rangesA));
+		ASSERT(noEmptyRangeInRanges(rangesB));
+	}
+	KeyRange emptyRange;
+	if (rangesA.empty() && rangesB.empty()) { // no mismatch
+		return Optional<std::pair<KeyRange, KeyRange>>();
+	} else if (rangesA.empty() && !rangesB.empty()) { // rangesA is empty while rangesB has a range
+		return std::make_pair(emptyRange, rangesB.front());
+	} else if (!rangesA.empty() && rangesB.empty()) { // rangesB is empty while rangesA has a range
+		return std::make_pair(rangesA.front(), emptyRange);
+	}
+	TraceEvent(SevVerbose, "RangesSameBeforeSort").detail("RangesA", rangesA).detail("Rangesb", rangesB);
+	// sort in ascending order
+	std::sort(rangesA.begin(), rangesA.end(), [](KeyRange a, KeyRange b) { return a.begin < b.begin; });
+	std::sort(rangesB.begin(), rangesB.end(), [](KeyRange a, KeyRange b) { return a.begin < b.begin; });
+	TraceEvent(SevVerbose, "RangesSameAfterSort").detail("RangesA", rangesA).detail("Rangesb", rangesB);
+	if (g_network->isSimulated()) {
+		ASSERT(elementsAreExclusiveWithEachOther(rangesA));
+		ASSERT(elementsAreExclusiveWithEachOther(rangesB));
+	}
+	if (rangesA.front().begin != rangesB.front().begin) { // rangeList heads mismatch
+		return std::make_pair(rangesA.front(), rangesB.front());
+	} else if (rangesA.back().end != rangesB.back().end) { // rangeList backs mismatch
+		return std::make_pair(rangesA.back(), rangesB.back());
+	}
+	int ia = 0;
+	int ib = 0;
+	KeyRangeRef rangeA = rangesA[0];
+	KeyRangeRef rangeB = rangesB[0];
+	KeyRange lastRangeA = Standalone(rangeA);
+	KeyRange lastRangeB = Standalone(rangeB);
+	while (true) {
+		if (rangeA.begin == rangeB.begin) {
+			if (rangeA.end == rangeB.end) {
+				if (rangeA.end == rangesA.back().end) {
+					break;
+				}
+				++ia;
+				++ib;
+				rangeA = rangesA[ia];
+				rangeB = rangesB[ib];
+				lastRangeA = Standalone(rangeA);
+				lastRangeB = Standalone(rangeB);
+			} else if (rangeA.end > rangeB.end) {
+				rangeA = KeyRangeRef(rangeB.end, rangeA.end);
+				++ib;
+				rangeB = rangesB[ib];
+				lastRangeB = Standalone(rangeB);
+			} else {
+				rangeB = KeyRangeRef(rangeA.end, rangeB.end);
+				++ia;
+				rangeA = rangesA[ia];
+				lastRangeA = Standalone(rangeA);
+			}
+		} else {
+			return std::make_pair(lastRangeA, lastRangeB);
+		}
+	}
+	return Optional<std::pair<KeyRange, KeyRange>>();
+}
+
+// Given an input server, get ranges within the input range via the input transaction
+// from the perspective of ServerKeys system key space
+// Input: (1) SS id; (2) transaction tr; (3) within range
+// Return AuditGetServerKeysRes, including: (1) complete range by a single read range;
+// (2) verison of the read; (3) ranges of the input SS
+ACTOR Future<AuditGetServerKeysRes> getThisServerKeysFromServerKeys(UID serverID, Transaction* tr, KeyRange range) {
+	state RangeResult readResult;
+	state AuditGetServerKeysRes res;
+
+	try {
+		wait(store(readResult,
+		           krmGetRanges(tr,
+		                        serverKeysPrefixFor(serverID),
+		                        range,
+		                        CLIENT_KNOBS->KRM_GET_RANGE_LIMIT,
+		                        CLIENT_KNOBS->KRM_GET_RANGE_LIMIT_BYTES)));
+		TraceEvent(SevVerbose, "AuditStorageGetThisServerKeysFromServerKeysReadDone", serverID)
+		    .detail("Range", range)
+		    .detail("Prefix", serverKeysPrefixFor(serverID))
+		    .detail("ResultSize", readResult.size())
+		    .detail("AduitServerID", serverID);
+
+		std::vector<KeyRange> ownRanges;
+		for (int i = 0; i < readResult.size() - 1; ++i) {
+			TraceEvent(SevVerbose, "AuditStorageGetThisServerKeysFromServerKeysAddToResult", serverID)
+			    .detail("ValueIsServerKeysFalse", readResult[i].value == serverKeysFalse)
+			    .detail("ServerHasKey", serverHasKey(readResult[i].value))
+			    .detail("Range", KeyRangeRef(readResult[i].key, readResult[i + 1].key))
+			    .detail("AduitServerID", serverID);
+			if (serverHasKey(readResult[i].value)) {
+				KeyRange shardRange;
+				ownRanges.push_back(Standalone(KeyRangeRef(readResult[i].key, readResult[i + 1].key)));
+			}
+		}
+		const KeyRange completeRange = Standalone(KeyRangeRef(range.begin, readResult.back().key));
+		Future<Version> getGrvF = tr->getReadVersion();
+		ASSERT(getGrvF.isReady());
+		const Version readAtVersion = getGrvF.get();
+		TraceEvent(SevVerbose, "AuditStorageGetThisServerKeysFromServerKeysEnd", serverID)
+		    .detail("AduitServerID", serverID)
+		    .detail("Range", range)
+		    .detail("Prefix", serverKeysPrefixFor(serverID))
+		    .detail("ReadAtVersion", readAtVersion)
+		    .detail("CompleteRange", completeRange)
+		    .detail("ResultSize", ownRanges.size());
+		res = AuditGetServerKeysRes(completeRange, readAtVersion, serverID, ownRanges, readResult.logicalSize());
+
+	} catch (Error& e) {
+		TraceEvent(SevDebug, "AuditStorageGetThisServerKeysError", serverID)
+		    .errorUnsuppressed(e)
+		    .detail("AduitServerID", serverID);
+		throw e;
+	}
+
+	return res;
+}
+
+// Given an input server, get ranges within the input range via the input transaction
+// from the perspective of KeyServers system key space
+// Input: (1) Audit Server ID (for logging); (2) transaction tr; (3) within range
+// Return AuditGetKeyServersRes, including : (1) complete range by a single read range; (2) verison of the read;
+// (3) map between SSes and their ranges --- in KeyServers space, a range corresponds to multiple SSes
+ACTOR Future<AuditGetKeyServersRes> getShardMapFromKeyServers(UID auditServerId, Transaction* tr, KeyRange range) {
+	state AuditGetKeyServersRes res;
+	state std::vector<Future<Void>> actors;
+	state RangeResult readResult;
+	state RangeResult UIDtoTagMap;
+	state int64_t totalShardsCount = 0;
+	state int64_t shardsInAnonymousPhysicalShardCount = 0;
+
+	try {
+		// read
+		actors.push_back(store(readResult,
+		                       krmGetRanges(tr,
+		                                    keyServersPrefix,
+		                                    range,
+		                                    CLIENT_KNOBS->KRM_GET_RANGE_LIMIT,
+		                                    CLIENT_KNOBS->KRM_GET_RANGE_LIMIT_BYTES)));
+		actors.push_back(store(UIDtoTagMap, tr->getRange(serverTagKeys, CLIENT_KNOBS->TOO_MANY)));
+		wait(waitForAll(actors));
+		ASSERT(!UIDtoTagMap.more && UIDtoTagMap.size() < CLIENT_KNOBS->TOO_MANY);
+		TraceEvent(SevVerbose, "AuditStorageGetThisServerKeysFromKeyServersReadDone", auditServerId)
+		    .detail("Range", range)
+		    .detail("ResultSize", readResult.size())
+		    .detail("AduitServerID", auditServerId);
+
+		// produce result
+		std::unordered_map<UID, std::vector<KeyRange>> serverOwnRanges;
+		for (int i = 0; i < readResult.size() - 1; ++i) {
+			std::vector<UID> src;
+			std::vector<UID> dest;
+			UID srcID;
+			UID destID;
+			decodeKeyServersValue(UIDtoTagMap, readResult[i].value, src, dest, srcID, destID);
+			if (srcID == anonymousShardId) {
+				shardsInAnonymousPhysicalShardCount++;
+			}
+			totalShardsCount++;
+			std::vector<UID> servers(src.size() + dest.size());
+			std::merge(src.begin(), src.end(), dest.begin(), dest.end(), servers.begin());
+			for (auto& ssid : servers) { // how to decide servers
+				serverOwnRanges[ssid].push_back(Standalone(KeyRangeRef(readResult[i].key, readResult[i + 1].key)));
+			}
+		}
+		const KeyRange completeRange = Standalone(KeyRangeRef(range.begin, readResult.back().key));
+		Future<Version> getGrvF = tr->getReadVersion();
+		ASSERT(getGrvF.isReady());
+		const Version readAtVersion = getGrvF.get();
+		TraceEvent(SevInfo, "AuditStorageGetThisServerKeysFromKeyServersEnd", auditServerId)
+		    .detail("Range", range)
+		    .detail("CompleteRange", completeRange)
+		    .detail("AtVersion", readAtVersion)
+		    .detail("ShardsInAnonymousPhysicalShardCount", shardsInAnonymousPhysicalShardCount)
+		    .detail("TotalShardsCount", totalShardsCount);
+		res = AuditGetKeyServersRes(completeRange, readAtVersion, serverOwnRanges, readResult.logicalSize());
+
+	} catch (Error& e) {
+		TraceEvent(SevDebug, "AuditStorageGetThisServerKeysFromKeyServersError", auditServerId)
+		    .errorUnsuppressed(e)
+		    .detail("AuditServerId", auditServerId);
+		throw e;
+	}
+
+	return res;
+}
+
 void clearAuditProgressMetadata(Transaction* tr, AuditType auditType, UID auditId) {
 	// There are two possible places to store AuditProgressMetadata:
 	// (1) auditServerBasedProgressRangeFor or (2) auditRangeBasedProgressRangeFor
