+++ fdbclient/include/fdbclient/ServerKnobs.h
@@ -877,6 +877,8 @@ public:
 	int PERSIST_FINISH_AUDIT_COUNT; // Num of persist complete/failed audits for each type
 	int AUDIT_RETRY_COUNT_MAX;
 	int CONCURRENT_AUDIT_TASK_COUNT_MAX;
+	bool AUDIT_DATAMOVE_PRE_CHECK;
+	bool AUDIT_DATAMOVE_POST_CHECK;
 	int BUGGIFY_BLOCK_BYTES;
 	int64_t STORAGE_RECOVERY_VERSION_LAG_LIMIT;
 	double STORAGE_DURABILITY_LAG_REJECT_THRESHOLD;
